chore(lint): restore the umbrella gate — clear 28 pre-existing findings, repair the rotted enrollment guard
make lint had been red at HEAD on this host behind two shadowed
failures, so its downstream guards had not run for some time:

- 28 pre-existing golangci-lint findings (pinned v2.12.2), none in
  files this loop's work touches: misspellings, redefined-builtin
  locals, unchecked tx.Rollback in tests, ctx-not-first test helpers
  (signatures flipped with all call sites), unused integration-test
  helpers (deleted), ineffectual assignment, De Morgan/switch/unlambda
  simplifications, and two Index()-can-return--1 guards in a helm
  contract test. Behavior preserved throughout; full unit suite green.

- scripts/check_enrollment_scoping.sh demanded wiring names deleted by
  62477bc (fix(tenancy): route pretenant credentials into silos), so
  the gate could never pass again: the exact-hash bare-pool resolve
  moved into internal/store/credentiallocators.go
  (pretenant_resolve_credential*), and enrollment mutations now run
  under tenancy.InTenant RLS — stronger than the shape the guard
  checked. The guard now asserts the current property: zero bare-pool
  calls in enrollment.go outside the deployment-scoped AgentCA
  repository (agent_ca has no tenant_id — migration 0041), bare-pool
  in the locators only inside the two SECURITY DEFINER resolvers, and
  the provider deny-list read intact. Its census also covers ALL
  receiver names (the old e.pool|a.pool pattern silently skipped
  c.pool), and the self-test now drives the same census function the
  live scan uses, planting both violation shapes.

Foundation-Loop: T-8cf3c4ec

+++ internal/tenantlife/tenantlife.go
@@ -73,7 +73,7 @@ const (
 	retentionPolicyAuditAction  = "lifecycle.retention_set"
 	maxRetentionAuditActorBytes = 256
 	auditRetentionDay           = 24 * time.Hour
-	maxAuditRetentionDays       = int64((1<<63 - 1) / int64(auditRetentionDay))
+	maxAuditRetentionDays       = (1<<63 - 1) / int64(auditRetentionDay)
 )
 
 var errRetentionAuditUnavailable = errors.New("tenantlife: retention audit sink is unavailable")
@@ -826,7 +826,7 @@ func (e *Engine) recordIRAttributionFailure(
 	ctx context.Context,
 	tenantID, actor, planID, failure string,
 ) error {
-	// A store or keyring may return after consuming/cancelling the caller's
+	// A store or keyring may return after consuming/canceling the caller's
 	// request context. Preserve values but give the forensic failure receipt a
 	// separate, finite window.
 	auditCtx, cancel := context.WithTimeout(context.WithoutCancel(ctx), irLifecycleReceiptTimeout)
